Finished sorting, delete confirmation

+++ script.js
@@ -4,6 +4,7 @@
 // ####################### GLOBAL VARIABLES ######################
 // ###############################################################
 let myLibrary = [];
+let currentlySortedBy;
 
 // ###############################################################
 // #################### HTML ELEMENT SELECTORS ###################
@@ -44,7 +45,7 @@ function addBookToLibrary() {
   }
 
   // Add the book to the myLibrary array
-  myLibrary.push(new Book(title.value, author.value, year.value, isRead.checked));
+  myLibrary.push(new Book(title.value, author.value, Number(year.value), isRead.checked));
 
   // Reset inputs
   title.value = author.value = year.value = "";
@@ -66,24 +67,18 @@ function displayBooks() {
       "beforeend",
       `
     <div class="book-item-container" data-id="${i}">
-        <div class="book-property-container">
-          <div class="label">Title:</div>
-          <div class="title-value">${book.title}</div>
+      <div class="title-value">${book.title}</div>
+      <div class="author-value">${book.author}</div>
+      <div class="year-value">${book.year}</div>
+      <label class="toggler-wrapper">
+        <input type="checkbox" id="read-checkbox" ${book.isRead ? "checked" : ""}>
+        <div class="toggler-slider">
+          <div class="toggler-knob"></div>
         </div>
-        <div class="book-property-container">
-          <div class="label">Author:</div>
-          <div class="author-value">${book.author}</div>
-        </div>
-        <div class="book-property-container">
-          <div class="label">Year:</div>
-          <div class="year-value">${book.year}</div>
-        </div>
-        <div class="book-property-container">
-          <div class="label">Read:</div>
-          <input type="checkbox" id="read-checkbox" ${book.isRead ? "checked" : ""}/>
-        </div>
-        <button id="delete-btn">Delete</button>
-      </div>
+      </label>
+      <button id="edit-btn">Edit</button>
+      <button id="delete-btn">Delete</button>
+    </div>
     `
     );
   });
@@ -95,6 +90,7 @@ function updateReadStatus(id, status) {
 }
 
 function deleteEntry(id) {
+  if (!confirm("Are you sure you want to delete this book from the list?")) return;
   myLibrary.splice(id, 1);
   saveToLocalStorage();
   displayBooks();
@@ -120,15 +116,15 @@ $containerAddBook.addEventListener("keypress", function (e) {
 
 $containerBookList.addEventListener("click", function (e) {
   const id = e.target.closest(".book-item-container").dataset.id;
+
+  // Listen for clicks on the "read" checkboxes
   if (e.target.id === "read-checkbox") {
     const status = e.target.checked;
     updateReadStatus(id, status);
   }
 
-  if (e.target.id === "delete-btn") {
-    // console.log("delte");
-    deleteEntry(id);
-  }
+  // Listen for "Delete" button presses
+  if (e.target.id === "delete-btn") deleteEntry(id);
 });
 
 // ###############################################################
@@ -136,3 +132,93 @@ $containerBookList.addEventListener("click", function (e) {
 // ###############################################################
 
 init();
+
+// TEMP
+
+const $titleHeader = document.querySelector(".title-header");
+const $authorHeader = document.querySelector(".author-header");
+const $yearHeader = document.querySelector(".year-header");
+const $readStatusHeader = document.querySelector(".read-status-header");
+
+$titleHeader.addEventListener("click", function () {
+  if (currentlySortedBy !== "title-asc") {
+    myLibrary.sort((a, b) => a.title.localeCompare(b.title));
+    currentlySortedBy = "title-asc";
+  } else if (currentlySortedBy === "title-asc") {
+    myLibrary.sort((a, b) => b.title.localeCompare(a.title));
+    currentlySortedBy = "title-desc";
+  }
+  displayBooks();
+});
+
+$authorHeader.addEventListener("click", function () {
+  if (currentlySortedBy !== "author-asc") {
+    myLibrary.sort((a, b) => {
+      // a.author.localeCompare(b.author);
+
+      if (a.author > b.author) {
+        return 1;
+      } else if (a.author < b.author) {
+        return -1;
+      }
+
+      if (a.title < b.title) {
+        return -1;
+      } else if (a.title > b.title) {
+        return 1;
+      } else {
+        return 0;
+      }
+    });
+    currentlySortedBy = "author-asc";
+  } else if (currentlySortedBy === "author-asc") {
+    myLibrary.sort((a, b) => b.author.localeCompare(a.author));
+    currentlySortedBy = "author-desc";
+  }
+  displayBooks();
+});
+
+$yearHeader.addEventListener("click", function () {
+  if (currentlySortedBy !== "year-asc") {
+    myLibrary.sort((a, b) => a.year - b.year);
+    currentlySortedBy = "year-asc";
+  } else if (currentlySortedBy === "year-asc") {
+    myLibrary.sort((a, b) => b.year - a.year);
+    currentlySortedBy = "year-desc";
+  }
+  displayBooks();
+});
+
+$readStatusHeader.addEventListener("click", function () {
+  if (currentlySortedBy !== "read-status-asc") {
+    myLibrary.sort((a, b) => {
+      // First, sort by the read status
+      if (a.isRead > b.isRead) {
+        return 1;
+      } else if (a.isRead < b.isRead) {
+        return -1;
+      }
+
+      // Then sort by author name
+      if (a.author < b.author) {
+        return -1;
+      } else if (a.author > b.author) {
+        return 1;
+      }
+
+      // Finally sort by title
+      if (a.title < b.title) {
+        return -1;
+      } else if (a.title > b.title) {
+        return 1;
+      } else {
+        return 0;
+      }
+    });
+    currentlySortedBy = "read-status-asc";
+  } else if (currentlySortedBy === "read-status-asc") {
+    myLibrary.sort((a, b) => b.isRead - a.isRead);
+    currentlySortedBy = "read-status-desc";
+  }
+  displayBooks();
+});
